requirements.txt
# Core Infrastructure
qdrant-client>=1.7.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
numpy>=1.24.0

# RAG & Embeddings
sentence-transformers>=2.2.2
transformers>=4.37.0
torch>=2.1.0
einops>=0.7.0
rank-bm25>=0.2.2

# Parsing & Chunking
tree-sitter>=0.20.0
# tree-sitter-languages>=1.10.0 # Incompatible with Python 3.13

# Data & Evaluation
datasets>=2.16.0
pandas>=2.0.0
Pillow>=10.0.0
openai>=1.10.0
tiktoken>=0.5.0
swebench>=2.0.0

# Visualization & Reporting
matplotlib>=3.8.0
seaborn>=0.13.0
tabulate>=0.9.0

# Development and testing (optional)
# pytest>=7.4.0
# jupyter>=1.0.0

# vLLM for LLM inference
vllm<0.11
# Development
pytest
questionary>=2.0.0
